# swagger.yaml
openapi: 3.0.0
info:
  title: Writemi
  contact: {}
  version: '1.0'
servers:
  - url: https://writemi.onrender.com
    variables: {}
paths:
...
  /auth/forgotPassword:
    post:
      tags:
        - Auth
      summary: Forgot Password
      operationId: ForgotPassword
      parameters: []
      requestBody:
        description: ''
        content:
          text/plain:
            schema:
              type: object
              example:
                recoveryKey: j5snanglbtdew7s
            example:
              recoveryKey: j5snanglbtdew7s
        required: true
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
      deprecated: false
      security: []
...
  /user/makeAdmin/{user_id}:
    patch:
      tags:
        - User
      summary: Make Admin
      operationId: MakeAdmin
      parameters:
        - name: user_id
          in: path
          description: ''
          required: true
          style: simple
          schema:
            type: string
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  msg: User updated to admin
              example:
                msg: User updated to admin
      deprecated: false
  /user/verifyUser/{user_id}:
    patch:
      tags:
        - User
      summary: Verify User
      operationId: VerifyUser
      parameters:
        - name: user_id
          in: path
          description: ''
          required: true
          style: simple
          schema:
            type: string
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  msg: User verification successful
              example:
                msg: User verification successful
      deprecated: false
  /user/banUser/{user_id}:
    patch:
      tags:
        - User
      summary: Ban User
      operationId: BanUser
      parameters:
        - name: user_id
          in: path
          description: ''
          required: true
          style: simple
          schema:
            type: string
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  msg: User banned successfully
              example:
                msg: User banned successfully
      deprecated: false
...
  /bookmark/addToBookmark:
    post:
      tags:
        - Bookmark
      summary: Add To Bookmark
      operationId: AddToBookmark
      parameters: []
      requestBody:
        description: ''
        content:
          text/plain:
            schema:
              type: object
              example:
                source: 6365acc5289536a28a41b693
                message: My name is maxwell
            example:
              source: 6365acc5289536a28a41b693
              message: My name is maxwell
        required: true
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  msg: Added to bookmark successfully
              example:
                msg: Added to bookmark successfully
      deprecated: false
...